+++ ldes-fragmentisers/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/fragmentation/FragmentationStrategyCreator.java
@@ -1,0 +1,7 @@
+package be.vlaanderen.informatievlaanderen.ldes.server.fragmentation;
+
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.entities.ViewSpecification;
+
+public interface FragmentationStrategyCreator {
+	FragmentationStrategy createFragmentationStrategyForView(ViewSpecification viewSpecification);
+}
