+++ ldes-fragmentisers/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/fragmentation/FragmentationStrategyCreatorImpl.java
@@ -1,0 +1,60 @@
+package be.vlaanderen.informatievlaanderen.ldes.server.fragmentation;
+
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.repository.LdesFragmentRepository;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.services.NonCriticalTasksExecutor;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.tree.member.repository.MemberRepository;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.entities.ViewSpecification;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.valueobjects.FragmentationConfig;
+import org.springframework.context.ApplicationContext;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+public class FragmentationStrategyCreatorImpl implements FragmentationStrategyCreator {
+	private final ApplicationContext applicationContext;
+	private final LdesFragmentRepository ldesFragmentRepository;
+	private final RootFragmentCreator rootFragmentCreator;
+	private final MemberRepository memberRepository;
+
+	public FragmentationStrategyCreatorImpl(ApplicationContext applicationContext,
+			LdesFragmentRepository ldesFragmentRepository,
+			RootFragmentCreator rootFragmentCreator,
+			MemberRepository memberRepository) {
+		this.applicationContext = applicationContext;
+		this.ldesFragmentRepository = ldesFragmentRepository;
+		this.rootFragmentCreator = rootFragmentCreator;
+		this.memberRepository = memberRepository;
+	}
+
+	public FragmentationStrategy createFragmentationStrategyForView(ViewSpecification viewSpecification) {
+		rootFragmentCreator.createRootFragmentForView(viewSpecification.getName());
+		NonCriticalTasksExecutor nonCriticalTasksExecutor = applicationContext.getBean(NonCriticalTasksExecutor.class);
+		FragmentationStrategy fragmentationStrategy = new FragmentationStrategyImpl(ldesFragmentRepository,
+				memberRepository, nonCriticalTasksExecutor);
+		if (viewSpecification.getFragmentations() != null) {
+			fragmentationStrategy = wrapFragmentationStrategy(viewSpecification.getFragmentations(),
+					fragmentationStrategy);
+		}
+		return fragmentationStrategy;
+	}
+
+	private FragmentationStrategy wrapFragmentationStrategy(List<FragmentationConfig> fragmentationConfigs,
+			FragmentationStrategy fragmentationStrategy) {
+		for (int i = fragmentationConfigs.size() - 1; i >= 0; i--) {
+			FragmentationConfig currentFragmentationConfig = fragmentationConfigs.get(i);
+			fragmentationStrategy = wrapFragmentationStrategyUsingFragmentationConfig(fragmentationStrategy,
+					currentFragmentationConfig);
+		}
+		return fragmentationStrategy;
+	}
+
+	private FragmentationStrategy wrapFragmentationStrategyUsingFragmentationConfig(
+			FragmentationStrategy fragmentationStrategy, FragmentationConfig currentFragmentationConfig) {
+		FragmentationStrategyWrapper fragmentationStrategyWrapper = (FragmentationStrategyWrapper) applicationContext
+				.getBean(currentFragmentationConfig.getName());
+		fragmentationStrategy = fragmentationStrategyWrapper.wrapFragmentationStrategy(applicationContext,
+				fragmentationStrategy, currentFragmentationConfig.getProperties());
+		return fragmentationStrategy;
+	}
+}
