+++ ldes-fragmentisers/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/fragmentation/FragmentationStrategyWrapper.java
@@ -1,0 +1,9 @@
+package be.vlaanderen.informatievlaanderen.ldes.server.fragmentation;
+
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.valueobjects.ConfigProperties;
+import org.springframework.context.ApplicationContext;
+
+public interface FragmentationStrategyWrapper {
+	FragmentationStrategy wrapFragmentationStrategy(ApplicationContext applicationContext,
+			FragmentationStrategy fragmentationStrategy, ConfigProperties fragmentationProperties);
+}
