--- ldes-queues/ldes-queue-in-memory/src/test/java/be/vlaanderen/informatievlaanderen/ldes/server/fragmentisers/mediator/InMemoryFragmentationMediatorTest.java
@@ -1,39 +1,0 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.fragmentisers.mediator;
-
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.services.FragmentationExecutor;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.tree.member.entities.Member;
-import org.awaitility.Awaitility;
-import org.awaitility.Durations;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.awaitility.Awaitility.await;
-import static org.mockito.Mockito.*;
-
-class InMemoryFragmentationMediatorTest {
-
-	private InMemoryFragmentationMediator fragmentationMediator;
-
-	private final FragmentationExecutor fragmentationExecutor = mock(FragmentationExecutor.class);
-
-	@BeforeEach
-	void setUp() {
-		fragmentationMediator = new InMemoryFragmentationMediator(fragmentationExecutor);
-	}
-
-	@Test
-	@DisplayName("Adding a member to the queue")
-	void when_MemberIsAddedForFragmentation_AThreadIsStartedWhichCallsTheFragmentationExecutor() {
-		Awaitility.reset();
-		Member member = mock(Member.class);
-		fragmentationMediator.addMemberToFragment(member);
-
-		await().pollDelay(Durations.ONE_MILLISECOND)
-				.atMost(Durations.TEN_SECONDS)
-				.until(() -> fragmentationMediator.ldesMembersToFragmentTracker.get() == 0);
-
-		verify(fragmentationExecutor, times(1)).executeFragmentation(member);
-	}
-
-}
