--- ldes-queues/pom.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <artifactId>ldes-server</artifactId>
-        <groupId>be.vlaanderen.informatievlaanderen.vsds</groupId>
-        <version>1.3.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>ldes-queues</artifactId>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>ldes-queue-none</module>
-        <module>ldes-queue-in-memory</module>
-    </modules>
-
-    <dependencies>
-        <!-- VSDS -->
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>ldes-server-domain</artifactId>
-        </dependency>
-        <!-- Testing -->
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
