--- ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/ldesfragment/services/FragmentationMediator.java
@@ -1,8 +1,0 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.services;
-
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.tree.member.entities.Member;
-
-public interface FragmentationMediator {
-
-	void addMemberToFragment(Member member);
-}
