--- ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/ldesfragment/services/FragmentationStrategyCollection.java
@@ -1,9 +1,0 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.services;
-
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.valueobjects.ViewName;
-
-import java.util.Map;
-
-public interface FragmentationStrategyCollection {
-	Map<ViewName, FragmentationStrategy> getFragmentationStrategyMap();
-}
