--- ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/ldesfragment/services/RefragmentationService.java
@@ -1,11 +1,0 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.services;
-
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.entities.LdesFragment;
-
-/**
- * The RefragmentationService creates fragments for a view based upon the
- * existing members.
- */
-public interface RefragmentationService {
-	void refragmentMembersForView(LdesFragment rootFragmentForView, FragmentationStrategy fragmentationStrategyForView);
-}
