--- ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/viewcreation/services/FragmentationStrategyCreator.java
@@ -1,8 +1,0 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.services;
-
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesfragment.services.FragmentationStrategy;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.viewcreation.entities.ViewSpecification;
-
-public interface FragmentationStrategyCreator {
-	FragmentationStrategy createFragmentationStrategyForView(ViewSpecification viewSpecification);
-}
